fjernet siste warning og flyttet hent organisasjoner et niva ut
